Updated entity enhancement
Updated the entity enhancement feature. When you move your mouse over an object, it will attempt to enhance it and put the enhanced image on the left side of the screen. The effect isn't perfect yet.

Currently, the mouse must be moving to enhance the image. Until I can find a way to further isolate the entities and hold onto a singular reference for them, I can't get a long-term reference. This is fine. The eye moves constantly, and those micro-movements are more than enough to assist the system, so it'll work either way.

I also started work on the pose estimator. This is going to end up being really important, as I would like to use it to help maintain the reference to the entity (i.e. how similar the position is to the entity's last recorded position), and eventually use it to help make the behavior on the screen more understandable.

As always, run the UI_tool.py and don't forget to install the requirements!