Closer to tool setup
Worked on the UI for the helper tool. Now, when you scan the image, it will show you what it found, as well as actually displaying the entity images on the side. It shows all categories.

It doesn't play yet, but quite honestly, this is so much easier to work with at this time.