Functionality Updates
The OD system now gives the entities some breathing room in their pictures. It looks a lot better. The video_main now has a specialized pop up window with the ability to display messages. It's drawing all of that directly into the canvas as it's doing it, and it still responds through hover, I love it. The sticky feature only works on the list items, though, unfortunately.

The main.py just kinda keeps it all together. It's looking great!